remove non-pro macro
