+++ macros/src/lib.rs
@@ -1,7 +1,5 @@
 use proc_macro2::TokenStream;
 
-/// TODO: migrate
-mod pro;
 mod testing;
 
 /// A macro to generate tests for Geo Engine services.
@@ -14,6 +12,11 @@ mod testing;
 /// - `test_execution` - `parallel` (default) or `serial`, which isolates this test from other tests
 /// - `before` - a function that is called before the context is created and the test is executed
 /// - `expect_panic` - if the test is expected to panic
+/// - `user` - `"admin"` if created session should be admin session
+/// - `quota_config` - a function that returns a [`crate::pro::util::config::Quota`] to use for the test
+/// - `oidc_db` - a tuple `(handle, f)` with
+///     - `handle` being a handle of an OpenID-Connect endpoint, preventing it from dropping too early, and
+///     - `f` begin function that returns a [`crate::pro::users::OidcManager`] to use for the test
 ///
 #[proc_macro_attribute]
 pub fn test(
@@ -21,36 +24,6 @@ pub fn test(
     item: proc_macro::TokenStream,
 ) -> proc_macro::TokenStream {
     match testing::test(attr.into(), &item.clone().into()) {
-        Ok(ts) => ts.into(),
-        Err(e) => token_stream_with_error(item.into(), e).into(),
-    }
-}
-
-/// A macro to generate tests for Geo Engine services.
-/// It automatically spins up a database context.
-///
-/// # Parameters
-///
-/// - `tiling_spec` - a function that returns a [`geoengine_datatypes::raster::TilingSpecification`] to use for the test
-/// - `query_ctx_chunk_size` - a function that returns a [`geoengine_operators::engine::ChunkByteSize`] to use for the test
-/// - `test_execution` - `parallel` (default) or `serial`, which isolates this test from other tests
-/// - `before` - a function that is called before the context is created and the test is executed
-/// - `expect_panic` - if the test is expected to panic
-/// - `user` - `"admin"` if created session should be admin session
-///
-/// ## Pro Parameters
-///
-/// - `quota_config` - a function that returns a [`crate::pro::util::config::Quota`] to use for the test
-/// - `oidc_db` - a tuple `(handle, f)` with
-///     - `handle` being a handle of an OpenID-Connect endpoint, preventing it from dropping too early, and
-///     - `f` begin function that returns a [`crate::pro::users::OidcManager`] to use for the test
-///
-#[proc_macro_attribute]
-pub fn pro_test(
-    attr: proc_macro::TokenStream,
-    item: proc_macro::TokenStream,
-) -> proc_macro::TokenStream {
-    match pro::test(attr.into(), &item.clone().into()) {
         Ok(ts) => ts.into(),
         Err(e) => token_stream_with_error(item.into(), e).into(),
     }
