+++ services/src/api/handlers/spatial_references.rs
@@ -269,7 +269,7 @@ mod tests {
     use super::*;
     use crate::contexts::Session;
     use crate::pro::contexts::ProPostgresContext;
-    use crate::pro::ge_context;
+    use crate::ge_context;
     use crate::pro::users::UserAuth;
     use crate::util::tests::send_test_request;
     use actix_web;
