+++ services/src/api/handlers/upload.rs
@@ -200,8 +200,8 @@ async fn list_upload_file_layers_handler(
 mod tests {
     use super::*;
     use crate::contexts::Session;
+    use crate::ge_context;
     use crate::pro::contexts::ProPostgresContext;
-    use crate::pro::ge_context;
     use crate::pro::users::UserAuth;
     use crate::util::tests::{send_test_request, SetMultipartBody, TestDataUploads};
     use actix_web::{http::header, test};
