+++ services/src/api/handlers/wms.rs
@@ -486,8 +486,8 @@ mod tests {
     use crate::datasets::listing::DatasetProvider;
     use crate::datasets::storage::DatasetStore;
     use crate::datasets::DatasetName;
+    use crate::ge_context;
     use crate::pro::contexts::ProPostgresContext;
-    use crate::pro::ge_context;
     use crate::pro::users::UserAuth;
     use crate::pro::util::tests::{admin_login, register_ndvi_workflow_helper};
     use crate::util::tests::{
