+++ services/src/lib.rs
@@ -19,6 +19,6 @@ pub mod pro;
 pub use geoengine_datatypes::test_data;
 
 // re-export test macro
-// pub mod ge_context {
-//     pub use geoengine_macros::test;
-// }
+pub mod ge_context {
+    pub use geoengine_macros::test;
+}
