+++ services/src/pro/contexts/postgres.rs
@@ -440,7 +440,7 @@ mod tests {
         LayerDb, LayerProviderDb, LayerProviderListing, LayerProviderListingOptions,
         INTERNAL_PROVIDER_ID,
     };
-    use crate::pro::ge_context;
+    use crate::ge_context;
     use crate::pro::permissions::{Permission, PermissionDb, Role, RoleDescription, RoleId};
     use crate::pro::users::{OidcTokens, SessionTokenStore};
     use crate::pro::users::{
