+++ services/src/pro/mod.rs
@@ -11,8 +11,3 @@ pub mod tasks;
 pub mod users;
 pub mod util;
 pub mod workflows;
-
-// re-export test macro
-pub mod ge_context {
-    pub use geoengine_macros::pro_test as test;
-}
