+++ services/src/pro/quota/mod.rs
@@ -222,9 +222,9 @@ mod tests {
     use super::*;
     use crate::{
         contexts::{ApplicationContext, SessionContext},
+        ge_context,
         pro::{
             contexts::ProPostgresContext,
-            ge_context,
             users::{UserAuth, UserCredentials, UserRegistration},
             util::tests::{admin_login, MockQuotaTracking},
         },
