+++ services/src/workflows/raster_stream.rs
@@ -172,7 +172,7 @@ fn send_result(
 mod tests {
     use super::*;
     use crate::pro::contexts::ProPostgresContext;
-    use crate::pro::ge_context;
+    use crate::ge_context;
     use crate::pro::users::UserAuth;
     use crate::util::tests::register_ndvi_workflow_helper;
     use crate::{contexts::ApplicationContext, pro::contexts::PostgresSessionContext};
